🔨(dev) enable individual configuration files out of tree files
In the `env.d/development` folder, we create `.dist` files used
as a template for each configuration files. The ones used in the
project are now generated.
Each developer will now be able to modify any key without risking
committing it to the tree files.
This implied deleting current configuration files present in the
distribution.